change .h to .hpp
