+++ core/testTask.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <iostream>
-#include "task.h"
+#include "task.hpp"
 
 class testTask :
 	public R_ATX::_task
